migrating to mocha+chai to get ordered tests and proper bail support.
