+++ src/tests/utils.ts
@@ -1,3 +1,3 @@
 export function formatTestName(step: number, testCaseIndex: number, testName: string, input: string, expected: number) {
-    return `Step ${step}.${testCaseIndex + 1} - ${testName}. input: ${input}; expected: ${expected}.`;
+    return `Step ${step}.${testCaseIndex + 1} - ${testName}.\n\tinput: "${input}"; expected: ${expected}`;
 }
